+++ Realization/backend/core/core-impl/src/main/java/eu/bcvsolutions/idm/core/workflow/service/impl/DefaultWorkflowProcessInstanceService.java
@@ -1,5 +1,6 @@
 package eu.bcvsolutions.idm.core.workflow.service.impl;
 
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashSet;
@@ -23,7 +24,14 @@ import org.activiti.engine.runtime.ProcessInstanceBuilder;
 import org.activiti.engine.runtime.ProcessInstanceQuery;
 import org.activiti.engine.task.IdentityLinkType;
 import org.activiti.engine.task.Task;
+import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.domain.Sort.Direction;
+import org.springframework.data.domain.Sort.Order;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.Assert;
@@ -33,8 +41,11 @@ import com.google.common.collect.ImmutableMap;
 import eu.bcvsolutions.idm.core.api.domain.CoreResultCode;
 import eu.bcvsolutions.idm.core.api.dto.IdmIdentityDto;
 import eu.bcvsolutions.idm.core.api.exception.ResultCodeException;
+import eu.bcvsolutions.idm.core.api.rest.domain.ResourcePage;
 import eu.bcvsolutions.idm.core.api.rest.domain.ResourcesWrapper;
 import eu.bcvsolutions.idm.core.api.service.IdmIdentityService;
+import eu.bcvsolutions.idm.core.rest.AbstractBaseDtoService;
+import eu.bcvsolutions.idm.core.security.api.domain.BasePermission;
 import eu.bcvsolutions.idm.core.security.api.service.SecurityService;
 import eu.bcvsolutions.idm.core.workflow.model.dto.WorkflowFilterDto;
 import eu.bcvsolutions.idm.core.workflow.model.dto.WorkflowProcessInstanceDto;
@@ -48,7 +59,7 @@ import eu.bcvsolutions.idm.core.workflow.service.WorkflowProcessInstanceService;
  *
  */
 @Service
-public class DefaultWorkflowProcessInstanceService implements WorkflowProcessInstanceService {
+public class DefaultWorkflowProcessInstanceService extends AbstractBaseDtoService<WorkflowProcessInstanceDto, WorkflowFilterDto> implements WorkflowProcessInstanceService {
 
 	@Autowired
 	private RuntimeService runtimeService;
@@ -73,12 +84,12 @@ public class DefaultWorkflowProcessInstanceService implements WorkflowProcessIns
 	public ProcessInstance startProcess(String definitionKey, String objectType, String applicant,
 			String objectIdentifier, Map<String, Object> variables) {
 		Assert.hasText(definitionKey, "Definition key cannot be null!");
-
+		UUID implementerId = securityService.getCurrentId();
+ 
 		IdmIdentityDto applicantIdentity = null;
 		if (applicant != null) {
 			applicantIdentity = identityService.getByUsername(applicant);
 		}
-		UUID implementerId = securityService.getCurrentId();
 		ProcessInstanceBuilder builder = runtimeService.createProcessInstanceBuilder()
 				.processDefinitionKey(definitionKey)//
 				.addVariable(WorkflowProcessInstanceService.OBJECT_TYPE, objectType)
@@ -96,12 +107,69 @@ public class DefaultWorkflowProcessInstanceService implements WorkflowProcessIns
 
 		ProcessInstance instance = builder.start();
 		if(!instance.isEnded()){
-			// Set applicant as owner of process
-			runtimeService.addUserIdentityLink(instance.getId(), applicant, IdentityLinkType.OWNER);
-			// Set current logged user (implementer) as starter of process
-			runtimeService.addUserIdentityLink(instance.getId(), securityService.getUsername(), IdentityLinkType.STARTER);
+			// must explicit check null, else throw org.activiti.engine.ActivitiIllegalArgumentException: userId and groupId cannot both be null
+			if (applicantIdentity != null) {
+				// Set applicant as owner of process
+				runtimeService.addUserIdentityLink(instance.getId(), applicantIdentity.getId().toString(), IdentityLinkType.OWNER);
+			}
+			if (implementerId != null) {
+				// Set current logged user (implementer) as starter of process
+				runtimeService.addUserIdentityLink(instance.getId(), implementerId.toString(), IdentityLinkType.STARTER);
+			}
 		}
 		return instance;
+	}
+	
+	@Override
+	public WorkflowProcessInstanceDto get(Serializable id, BasePermission... permission) {
+		Assert.notNull(id);
+		return this.get(String.valueOf(id));
+	}
+	
+	@Override
+	public Page<WorkflowProcessInstanceDto> find(WorkflowFilterDto filter, Pageable pageable,
+			BasePermission... permission) {
+		// we must call original method search because is there check flag checkRight
+		if (pageable != null) {
+			filter.setPageNumber(pageable.getPageNumber());
+			filter.setPageSize(pageable.getPageSize());
+			//
+			String fieldForSort = null;
+			boolean ascSort = false;
+			boolean descSort = false;
+			if (pageable != null) {
+				Sort sort = pageable.getSort();
+				if (sort != null) {
+					for (Order order : sort) {
+						if (!StringUtils.isEmpty(order.getProperty())) {
+							// TODO: now is implemented only one property sort 
+							fieldForSort = order.getProperty();
+							if (order.getDirection() == Direction.ASC) {
+								ascSort = true;
+							} else if (order.getDirection() == Direction.DESC) {
+								descSort = true;
+							}
+							break;
+						}
+						
+					}
+				}
+			}
+			filter.setSortAsc(ascSort);
+			filter.setSortDesc(descSort);
+			filter.setSortByFields(fieldForSort);
+		}
+		ResourcesWrapper<WorkflowProcessInstanceDto> search = this.search(filter);
+		//
+		ResourcePage pages = search.getPage();
+		List<WorkflowProcessInstanceDto> processes = (List<WorkflowProcessInstanceDto>) search.getResources();
+		//
+		return new PageImpl<WorkflowProcessInstanceDto>(processes, pageable, pages.getTotalPages());
+	}
+	
+	@Override
+	public Page<WorkflowProcessInstanceDto> find(Pageable pageable, BasePermission... permission) {
+		return this.find(new WorkflowFilterDto(), pageable, permission);
 	}
 
 	@Override
@@ -190,6 +258,35 @@ public class DefaultWorkflowProcessInstanceService implements WorkflowProcessIns
 		Collection<WorkflowProcessInstanceDto> resources = this.search(filter).getResources();
 		return !resources.isEmpty() ? resources.iterator().next() : null;
 	}
+	
+	@Override
+	public WorkflowProcessInstanceDto get(String processInstanceId, boolean checkRight) {
+		WorkflowFilterDto filter = new WorkflowFilterDto();
+		filter.setProcessInstanceId(processInstanceId);
+		filter.setSortAsc(true);
+		Collection<WorkflowProcessInstanceDto> resources = this.searchInternal(filter, checkRight).getResources();
+		return !resources.isEmpty() ? resources.iterator().next() : null;
+	}
+	
+	@Override
+	public void delete(WorkflowProcessInstanceDto dto, BasePermission... permission) {
+		this.delete(dto.getId(), null);
+	}
+	
+	@Override
+	public void deleteById(Serializable id, BasePermission... permission) {
+		this.delete(String.valueOf(id), null);
+	}
+	
+	@Override
+	public void deleteInternalById(Serializable id) {
+		this.delete(String.valueOf(id), null);
+	}
+	
+	@Override
+	public void deleteInternal(WorkflowProcessInstanceDto dto) {
+		this.delete(dto.getId(), null);
+	}
 
 	@Override
 	public WorkflowProcessInstanceDto delete(String processInstanceId, String deleteReason) {
@@ -274,5 +371,4 @@ public class DefaultWorkflowProcessInstanceService implements WorkflowProcessIns
 		
 		return dto;
 	}
-
 }