+++ Realization/backend/core/core-impl/src/test/java/eu/bcvsolutions/idm/core/workflow/permissions/ChangeIdentityPermissionTest.java
@@ -11,6 +11,7 @@ import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
 import org.springframework.transaction.annotation.Transactional;
 
 import eu.bcvsolutions.idm.InitTestData;
@@ -21,6 +22,7 @@ import eu.bcvsolutions.idm.core.api.domain.RoleRequestedByType;
 import eu.bcvsolutions.idm.core.api.dto.IdmConceptRoleRequestDto;
 import eu.bcvsolutions.idm.core.api.dto.IdmIdentityContractDto;
 import eu.bcvsolutions.idm.core.api.dto.IdmIdentityDto;
+import eu.bcvsolutions.idm.core.api.dto.IdmIdentityRoleDto;
 import eu.bcvsolutions.idm.core.api.dto.IdmRoleDto;
 import eu.bcvsolutions.idm.core.api.dto.IdmRoleGuaranteeDto;
 import eu.bcvsolutions.idm.core.api.dto.IdmRoleRequestDto;
@@ -30,8 +32,11 @@ import eu.bcvsolutions.idm.core.api.service.IdmIdentityContractService;
 import eu.bcvsolutions.idm.core.api.service.IdmIdentityService;
 import eu.bcvsolutions.idm.core.api.service.IdmRoleRequestService;
 import eu.bcvsolutions.idm.core.api.service.IdmRoleService;
+import eu.bcvsolutions.idm.core.security.api.service.SecurityService;
 import eu.bcvsolutions.idm.core.workflow.model.dto.WorkflowFilterDto;
+import eu.bcvsolutions.idm.core.workflow.model.dto.WorkflowHistoricProcessInstanceDto;
 import eu.bcvsolutions.idm.core.workflow.model.dto.WorkflowTaskInstanceDto;
+import eu.bcvsolutions.idm.core.workflow.service.WorkflowHistoricProcessInstanceService;
 import eu.bcvsolutions.idm.core.workflow.service.WorkflowTaskInstanceService;
 import eu.bcvsolutions.idm.test.api.TestHelper;
 
@@ -46,6 +51,7 @@ public class ChangeIdentityPermissionTest extends AbstractCoreWorkflowIntegratio
 
 	private static final String APPROVE_ROLE_BY_GUARANTEE_KEY = "approve-role-by-guarantee";
 	private static final String SECURITY_ROLE_TEST = "securityRoleTest";
+	private static final String APPROVE_BY_HELPDESK_ROLE = "idm.sec.core.wf.approval.helpdesk.role";
 	private static final String APPROVE_BY_SECURITY_ENABLE = "idm.sec.core.wf.approval.security.enabled";
 	private static final String APPROVE_BY_MANAGER_ENABLE = "idm.sec.core.wf.approval.manager.enabled";
 	private static final String APPROVE_BY_USERMANAGER_ENABLE = "idm.sec.core.wf.approval.usermanager.enabled";
@@ -54,12 +60,14 @@ public class ChangeIdentityPermissionTest extends AbstractCoreWorkflowIntegratio
 	//
 	@Autowired private TestHelper helper;
 	@Autowired private WorkflowTaskInstanceService workflowTaskInstanceService;
+	@Autowired private WorkflowHistoricProcessInstanceService workflowHistoricProcessInstanceService;
 	@Autowired private IdmIdentityService identityService;
 	@Autowired private IdmIdentityContractService identityContractService;
 	@Autowired private IdmConceptRoleRequestService conceptRoleRequestService;
 	@Autowired private IdmRoleRequestService roleRequestService;
 	@Autowired private IdmRoleService roleService;
 	@Autowired private IdmConfigurationService configurationService;
+	@Autowired private SecurityService securityService;
 
 	@Before
 	public void login() {
@@ -94,19 +102,24 @@ public class ChangeIdentityPermissionTest extends AbstractCoreWorkflowIntegratio
 		assertEquals(RoleRequestState.IN_PROGRESS, request.getState());
 		
 		WorkflowFilterDto taskFilter = new WorkflowFilterDto();
-		List<WorkflowTaskInstanceDto> tasks = (List<WorkflowTaskInstanceDto>) workflowTaskInstanceService.search(taskFilter).getResources();
+		taskFilter.setCandidateOrAssigned(securityService.getCurrentUsername());
+		List<WorkflowTaskInstanceDto> tasks = workflowTaskInstanceService.find(taskFilter, null).getContent();
 		assertEquals(0, tasks.size());
 		
 		loginAsAdmin(InitTestData.TEST_ADMIN_USERNAME);
 		// HELPDESK
+		taskFilter.setCandidateOrAssigned(InitTestData.TEST_ADMIN_USERNAME);
 		checkAndCompleteOneTask(taskFilter, InitTestData.TEST_USER_1, "approve");
 		// MANAGER
 		loginAsAdmin(InitTestData.TEST_USER_2);
+		taskFilter.setCandidateOrAssigned(InitTestData.TEST_USER_2);
 		checkAndCompleteOneTask(taskFilter, InitTestData.TEST_USER_1, "approve");
 		// USER MANAGER
 		loginAsAdmin(InitTestData.TEST_ADMIN_USERNAME);
+		taskFilter.setCandidateOrAssigned(InitTestData.TEST_ADMIN_USERNAME);
 		checkAndCompleteOneTask(taskFilter, InitTestData.TEST_USER_1, "approve");
 		// SECURITY
+		taskFilter.setCandidateOrAssigned(InitTestData.TEST_ADMIN_USERNAME);
 		checkAndCompleteOneTask(taskFilter, InitTestData.TEST_USER_1, "approve");
 		
 		request = roleRequestService.get(request.getId());
@@ -136,15 +149,18 @@ public class ChangeIdentityPermissionTest extends AbstractCoreWorkflowIntegratio
 		assertEquals(RoleRequestState.IN_PROGRESS, request.getState());
 		
 		WorkflowFilterDto taskFilter = new WorkflowFilterDto();
-		List<WorkflowTaskInstanceDto> tasks = (List<WorkflowTaskInstanceDto>) workflowTaskInstanceService.search(taskFilter).getResources();
+		taskFilter.setCandidateOrAssigned(securityService.getCurrentUsername());
+		List<WorkflowTaskInstanceDto> tasks = workflowTaskInstanceService.find(taskFilter, null).getContent();
 		assertEquals(0, tasks.size());
 		
 		// HELPDESK - must be skipped
 		// MANAGER
 		loginAsAdmin(InitTestData.TEST_USER_2);
+		taskFilter.setCandidateOrAssigned(InitTestData.TEST_USER_2);
 		checkAndCompleteOneTask(taskFilter, InitTestData.TEST_USER_1, "approve");
 		// USER MANAGER
 		loginAsAdmin(InitTestData.TEST_ADMIN_USERNAME);
+		taskFilter.setCandidateOrAssigned(InitTestData.TEST_ADMIN_USERNAME);
 		checkAndCompleteOneTask(taskFilter, InitTestData.TEST_USER_1, "approve");
 		// SECURITY - must be skipped
 		request = roleRequestService.get(request.getId());
@@ -174,12 +190,14 @@ public class ChangeIdentityPermissionTest extends AbstractCoreWorkflowIntegratio
 		assertEquals(RoleRequestState.IN_PROGRESS, request.getState());
 		
 		WorkflowFilterDto taskFilter = new WorkflowFilterDto();
-		List<WorkflowTaskInstanceDto> tasks = (List<WorkflowTaskInstanceDto>) workflowTaskInstanceService.search(taskFilter).getResources();
+		taskFilter.setCandidateOrAssigned(securityService.getCurrentUsername());
+		List<WorkflowTaskInstanceDto> tasks = workflowTaskInstanceService.find(taskFilter, null).getContent();
 		assertEquals(0, tasks.size());
 		
 		// HELPDESK - must be skipped
 		// MANAGER
 		loginAsAdmin(InitTestData.TEST_USER_2);
+		taskFilter.setCandidateOrAssigned(InitTestData.TEST_USER_2);
 		checkAndCompleteOneTask(taskFilter, InitTestData.TEST_USER_1, "disapprove");
 		request = roleRequestService.get(request.getId());
 		assertEquals(RoleRequestState.DISAPPROVED, request.getState());
@@ -221,22 +239,27 @@ public class ChangeIdentityPermissionTest extends AbstractCoreWorkflowIntegratio
 		assertEquals(RoleRequestState.IN_PROGRESS, request.getState());
 		
 		WorkflowFilterDto taskFilter = new WorkflowFilterDto();
-		List<WorkflowTaskInstanceDto> tasks = (List<WorkflowTaskInstanceDto>) workflowTaskInstanceService.search(taskFilter).getResources();
+		taskFilter.setCandidateOrAssigned(securityService.getCurrentUsername());
+		List<WorkflowTaskInstanceDto> tasks = workflowTaskInstanceService.find(taskFilter, null).getContent();
 		assertEquals(0, tasks.size());
 		
 		// HELPDESK - must be skipped
 		// MANAGER
 		loginAsAdmin(InitTestData.TEST_USER_2);
+		taskFilter.setCandidateOrAssigned(InitTestData.TEST_USER_2);
 		checkAndCompleteOneTask(taskFilter, InitTestData.TEST_USER_1, "approve");
 		// USER MANAGER
 		loginAsAdmin(InitTestData.TEST_ADMIN_USERNAME);
+		taskFilter.setCandidateOrAssigned(InitTestData.TEST_ADMIN_USERNAME);
 		checkAndCompleteOneTask(taskFilter, InitTestData.TEST_USER_1, "approve");
 		// Subprocess - approve by GUARANTEE
 		loginAsAdmin(InitTestData.TEST_USER_2);
+		taskFilter.setCandidateOrAssigned(InitTestData.TEST_USER_2);
 		checkAndCompleteOneTask(taskFilter, InitTestData.TEST_USER_1, "approve");
 
 		// SECURITY 
 		loginAsAdmin(InitTestData.TEST_ADMIN_USERNAME);
+		taskFilter.setCandidateOrAssigned(InitTestData.TEST_ADMIN_USERNAME);
 		checkAndCompleteOneTask(taskFilter, InitTestData.TEST_USER_1, "approve");
 		
 		request = roleRequestService.get(request.getId());
@@ -278,7 +301,8 @@ public class ChangeIdentityPermissionTest extends AbstractCoreWorkflowIntegratio
 		assertEquals(RoleRequestState.IN_PROGRESS, request.getState());
 		
 		WorkflowFilterDto taskFilter = new WorkflowFilterDto();
-		List<WorkflowTaskInstanceDto> tasks = (List<WorkflowTaskInstanceDto>) workflowTaskInstanceService.search(taskFilter).getResources();
+		taskFilter.setCandidateOrAssigned(securityService.getCurrentUsername());
+		List<WorkflowTaskInstanceDto> tasks = workflowTaskInstanceService.find(taskFilter, null).getContent();
 		assertEquals(0, tasks.size());
 
 		// HELPDESK 	turn off
@@ -286,6 +310,7 @@ public class ChangeIdentityPermissionTest extends AbstractCoreWorkflowIntegratio
 		// USER MANAGER	turn off
 		// SECURITY
 		loginAsAdmin(InitTestData.TEST_USER_1);
+		taskFilter.setCandidateOrAssigned(InitTestData.TEST_USER_1);
 		checkAndCompleteOneTask(taskFilter, InitTestData.TEST_USER_1, "approve");
 		
 		request = roleRequestService.get(request.getId());
@@ -295,6 +320,100 @@ public class ChangeIdentityPermissionTest extends AbstractCoreWorkflowIntegratio
 		assertNotNull(concept.getWfProcessId());
 	}
 
+	@Test
+	@Transactional
+	public void testTaskCount() {
+		configurationService.setValue(APPROVE_BY_SECURITY_ENABLE, "false");
+		configurationService.setValue(APPROVE_BY_MANAGER_ENABLE, "false");
+		configurationService.setValue(APPROVE_BY_HELPDESK_ENABLE, "true");
+		configurationService.setValue(APPROVE_BY_USERMANAGER_ENABLE, "false");
+		//
+		loginAsAdmin(InitTestData.TEST_ADMIN_USERNAME);
+		IdmIdentityDto test1 = helper.createIdentity();
+		IdmIdentityDto guarantee = helper.createIdentity();
+		
+		// Guarantee
+		int priority = 500;
+		IdmRoleDto role = helper.createRole();
+		role.setPriority(priority);		
+		IdmRoleGuaranteeDto roleGuarantee = new IdmRoleGuaranteeDto();
+		roleGuarantee.setRole(role.getId());
+		roleGuarantee.setGuarantee(guarantee.getId());
+		role.getGuarantees().add(roleGuarantee);
+		role = roleService.save(role);
+		// set approve by guarantee
+		configurationService.setValue(IdmRoleService.WF_BY_ROLE_PRIORITY_PREFIX + priority, APPROVE_ROLE_BY_GUARANTEE_KEY);
+
+		// helpdesk role and identity
+		IdmRoleDto helpdeskRole = helper.createRole();
+		IdmIdentityDto helpdeskIdentity = helper.createIdentity();
+		// add role dirctly
+		helper.createIdentityRole(helpdeskIdentity, helpdeskRole);
+		configurationService.setValue(APPROVE_BY_HELPDESK_ROLE, helpdeskRole.getCode());
+		
+		IdmIdentityContractDto contract = helper.getPrimeContract(test1.getId());
+		
+		// check task before create request
+		loginAsAdmin(test1.getUsername());
+		int taskCount = getHistoricProcess().size();
+		
+		IdmRoleRequestDto request = createRoleRequest(test1);
+		request = roleRequestService.save(request);
+		
+		IdmConceptRoleRequestDto concept = createRoleConcept(role, contract, request);
+		concept = conceptRoleRequestService.save(concept);
+		
+		roleRequestService.startRequestInternal(request.getId(), true);
+		request = roleRequestService.get(request.getId());
+		assertEquals(RoleRequestState.IN_PROGRESS, request.getState());
+		
+		WorkflowFilterDto taskFilter = new WorkflowFilterDto();
+		taskFilter.setCandidateOrAssigned(securityService.getCurrentUsername());
+		List<WorkflowTaskInstanceDto> tasks = workflowTaskInstanceService.find(taskFilter, null).getContent();
+		assertEquals(0, tasks.size());
+		
+		// check tasks after create request, must be +1
+		loginAsAdmin(test1.getUsername());
+		int taksCountAfter = getHistoricProcess().size();
+		assertEquals(taskCount + 1, taksCountAfter);
+		
+		// HELPDESK
+		loginAsAdmin(helpdeskIdentity.getUsername());
+		taskFilter.setCandidateOrAssigned(helpdeskIdentity.getUsername());
+		checkAndCompleteOneTask(taskFilter, test1.getUsername(), "approve");
+
+		// check tasks by identity, must be + 2 (main process + sub process)
+		loginAsAdmin(test1.getUsername());
+		taksCountAfter = getHistoricProcess().size();
+		assertEquals(taskCount + 2, taksCountAfter);
+		
+		// Subprocess - approve by GUARANTEE
+		loginAsAdmin(guarantee.getUsername());
+		taskFilter.setCandidateOrAssigned(guarantee.getUsername());
+		checkAndCompleteOneTask(taskFilter, test1.getUsername(), "approve");
+
+		request = roleRequestService.get(request.getId());
+		assertEquals(RoleRequestState.EXECUTED, request.getState());
+		assertNotNull(request.getWfProcessId());
+		concept = conceptRoleRequestService.get(concept.getId());
+		assertNotNull(concept.getWfProcessId());
+		
+		// check task on the end (same as before)
+		loginAsAdmin(test1.getUsername());
+		taksCountAfter = getHistoricProcess().size();
+		assertEquals(taskCount + 2, taksCountAfter);
+	}
+	
+	/**
+	 * Return {@link WorkflowHistoricProcessInstanceDto} for current logged user
+	 * 
+	 * @return
+	 */
+	private List<WorkflowHistoricProcessInstanceDto> getHistoricProcess() {
+		WorkflowFilterDto taskFilter = new WorkflowFilterDto();
+		taskFilter.setCandidateOrAssigned(securityService.getCurrentId().toString());
+		return workflowHistoricProcessInstanceService.find(taskFilter, null).getContent();
+	}
 
 	private IdmConceptRoleRequestDto createRoleConcept(IdmRoleDto adminRole, IdmIdentityContractDto contract,
 			IdmRoleRequestDto request) {