+++ Realization/frontend/czechidm-core/src/content/task/Task.js
@@ -27,6 +27,7 @@ class Task extends Basic.AbstractContent {
   componentDidMount() {
     this.selectNavigationItem('tasks');
     const { taskID } = this.props.params;
+    this.context.store.dispatch(workflowTaskInstanceManager.fetchPermissions(taskID, null));
     this.context.store.dispatch(workflowTaskInstanceManager.fetchEntityIfNeeded(taskID, null, (json, error) => {
       if (error) {
         // task isn't exists or is solved
@@ -60,7 +61,7 @@ class Task extends Basic.AbstractContent {
   }
 
   render() {
-    const { readOnly, task } = this.props;
+    const { readOnly, task, _permissions } = this.props;
     const { nonExistentTask } = this.state;
     let DetailComponent;
     if (task && task.formKey) {
@@ -94,7 +95,7 @@ class Task extends Basic.AbstractContent {
     return (
       <div>
         {task ?
-        <DetailComponent task={task} uiKey="dynamic-task-detail" taskManager={workflowTaskInstanceManager} readOnly={readOnly}/>
+        <DetailComponent task={task} uiKey="dynamic-task-detail" taskManager={workflowTaskInstanceManager} readOnly={readOnly} canExecute={workflowTaskInstanceManager.canExecute(task, _permissions)}/>
         :
         <Basic.Well showLoading/>
         }
@@ -105,20 +106,22 @@ class Task extends Basic.AbstractContent {
 
 Task.propTypes = {
   task: PropTypes.object,
-  readOnly: PropTypes.bool
+  readOnly: PropTypes.bool,
+  _permissions: PropTypes.arrayOf(PropTypes.string)
 };
 
 Task.defaultProps = {
   task: null,
-  readOnly: false
+  readOnly: false,
+  _permissions: null
 };
 
 function select(state, component) {
   const { taskID } = component.params;
   const task = workflowTaskInstanceManager.getEntity(state, taskID);
-
   return {
-    task
+    task,
+    _permissions: workflowTaskInstanceManager.getPermissions(state, null, taskID)
   };
 }