+++ Realization/frontend/czechidm-core/src/redux/data/WorkflowHistoricTaskInstanceManager.js
@@ -17,6 +17,6 @@ export default class WorkflowHistoricTaskInstanceManager extends EntityManager {
   }
 
   getCollectionType() {
-    return 'resources';
+    return 'workflowHistoricTaskInstances';
   }
 }