+++ CHANGELOG.md
@@ -7,6 +7,11 @@ All notable changes to this project will be documented in this file.
 
   - [#920](https://redmine.czechidm.com/issues/920) - Event processor - support conditional processing. New method ``EntityEventProcessor#conditional``.
   - [#922](https://redmine.czechidm.com/issues/922) - ``IdmIdentityService#findValidByRole`` method was added - return valid identities with assigned valid role (valid contract and identity role relation).
+  - Workflow task instances has permissions now. Add new permission for read and execute task.
+
+### Changed
+
+  - All workflow services and controllers now using DTO layer, for more information read [#529](https://redmine.czechidm.com/issues/529).
 
 ## [7.6.1]